Fixed broken page (TIDOC-2450)
At the end of the Map.yml file, there are two example. The last example
doesn’t have an “examples” header which was causing the valid yaml to
treat the last example as a method (which passes the validation test)
but results in rendering no content. With the additional insertion of
examples:﻿ on line 363 of Map.yml, the content shows up again.