fix(ios): use layoutMargins to apply padding (#253)
* Added layoutMargins property for iOS

Use map.view.layoutMargins to update the edge insets for map controls (eg. compass, scale, ...). Note that with iOS 11+ the window/view safe area is already included.

* Added layoutMargins

iOS is seemingly having an issue with setVisibleMapRect which is what the padding property/method uses to update the edgeInsets of map controls (eg. compass, scale, etc.). This is another method that works, however, there is currently no option to apply an animation.

* Update `padding`/remove `layoutMargins` for parity

Replace `setVisibleMapRect:edgePadding:animated:` with `setLayoutMargins` as the former is currently not working due to some know issue(s) with MapKit.

* Update `padding`/remove `layoutMargins` for parity

* Version bump

Updated the `padding` property/method to use `.layoutMargins`